upgrade codex-africanus (#729)
* correct codex-africanus version in setup.py

* Removed outdated dtype keyword when fitting spi's

+++ DDFacet/Imager/HOGBOM/ClassModelMachineHogbom.py
@@ -332,8 +332,7 @@ class ClassModelMachine(ClassModelMachinebase.ClassModelMachine):
             freqsDask = da.from_array(np.array(self.GridFreqs).astype(np.float64), chunks=(self.Nchan))
 
             alpha, varalpha, Iref, varIref = fit_spi_components(FitCubeDask, weightsDask,
-                                                                freqsDask, self.RefFreq,
-                                                                dtype=np.float64).compute()
+                                                                freqsDask, self.RefFreq).compute()
         except Exception as e:
             traceback_str = traceback.format_exc(e)
             print("Warning - Failed at importing africanus spi fitter. This could be an issue with the dask " \
